fix up confusion about max shards and shard count

+++ src/api_context.odin
@@ -295,7 +295,7 @@ ctx_type_config :: #force_inline proc(ctx: ^TinaContext) -> ^TypeDescriptor {
 	return &shard.type_descriptors[type_id]
 }
 
-ctx_shard_id :: #force_inline proc(ctx: ^TinaContext) -> u16 {
+ctx_shard_id :: #force_inline proc(ctx: ^TinaContext) -> u8 {
 	shard := _ctx_extract_shard(ctx)
 	return shard.id
 }
